summarizing_docker/app/requirements.txt
fastapi
uvicorn
pinecone-client
pypdf2
pandas
numpy
transformers
accelerate
huggingface_hub[cli]
bitsandbytes
torch